+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Extensions
 
 - `Collection.nonEmpty` property is added, that returns non-empty `Collection`, or `nil`.
 - `withContinuousObservationTracking(of:on:execute:)` method is added, that tracks access to properties in a `class`-type continuously
+- `View.inlineNavigationTitle(_:)` method is added with `LocalizedStringKey`
 
 Extra - XCode Templates